Big Revision to the entire codebase to allow compilation and basic rendering on new compilers (Focus on Mac OSX & clang) FIXED: OpenGL rendering of samples now shows the correct pointSize.

+++ Core/qcontour.h
@@ -35,7 +35,7 @@ public:
 public:
     bool bDrawColorbar;
     QColor plotColor;
-    int plotThickness;
+    float plotThickness;
     Qt::PenStyle style;
 };
